Safely cast jsonData to InternalMessageData
This commit logs a warning when a jsonObject is passed which does not conform to the InternalMessageData type, otherwise retaining the existing behaviour of defaulting to an empty object.

This assists users of the library who might pass an otherwise-Codable object in as data, and then observe that the value they pass is dropped, without changing any other behaviour.