Fix TabBarMinimize behaviour on visit
When we perform a visit we add several views to the
VisitableView in the order:

  1. activityIndicatorView
  2. screenshotContainerView
  3. webView

Whenever a visits starts we move the activityIndicator to the bottom
of the hierarchy (using `bringSubviewToFront`) and before displaying
the view we remove the screenshotContainerView (using
`removeFromSuperview`).

As a result the webView ends up being the first view in the hierarchy
when inspecting the application at any point after loading the
Visitable.

When starting the application and performing the initial load this
process causes no issues. However, with subsequent visits this results
in behaviour like the UINavigationBar.prefersLargeTitles or
TabBarMinimize functionality in iOS 26 not working as expected.  It
appears like scrolling through the webview is not detected, therefore
therefore not triggering these specific behaviours.

While this is undocumented by Apple I found several sources[^1][^2]
that indicate that behaviour like UINavigationBar.prefersLargeTitles
requires that the scollable view is the first child in the view
hierarchy.  While our webview is eventually the first view in the
hierachy, the fact that it isn't during loading seems to cause the
issues that I experienced.

This led me to experiment with insertSubview to ensure the webView is
always the first child in the view hierarchy.  As a result I have seen
no more issues with the behaviours mentioned above.

A better solution might be to rework the way we add views to the
VisitableView such that the webview always ends up in the initial
position.  While I attempted such a solution I found that using
insertSubview is more explicit and less likely to break by future
changes to this file.

[^1]: https://swiftsenpai.com/development/large-title-uinavigationbar-glitches/
[^2]: https://github.com/software-mansion/react-native-screens/issues/1034